Link sky_shell on linux statically against osmesa instead of runtime switch
